refactor(logo): move Logo to lib/app/components.tsx as named export
Moves the enterprise-aware Logo component from refresh-components/ to
lib/app/ alongside the rest of the app-level utilities. Updates all
consumers to the new named import and updates the Jest mock shape.
