+++ web/src/app/app/components/WelcomeMessage.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import Logo from "@/refresh-components/Logo";
+import { Logo } from "@/lib/app/components";
 import {
   getRandomGreeting,
   GREETING_MESSAGES,
