+++ web/src/app/auth/waiting-on-verification/page.tsx
@@ -4,7 +4,7 @@ import { AuthTypeMetadata } from "@/lib/auth/types";
 import { redirect } from "next/navigation";
 import { User } from "@/lib/types";
 import { RequestNewVerificationEmail } from "./RequestNewVerificationEmail";
-import Logo from "@/refresh-components/Logo";
+import { Logo } from "@/lib/app/components";
 import { Text } from "@opal/components";
 import { markdown } from "@opal/utils";
