+++ web/src/lib/app/components.tsx
@@ -19,12 +19,7 @@ export interface LogoProps {
   onyxBranded?: boolean;
 }
 
-export default function Logo({
-  folded,
-  size,
-  className,
-  onyxBranded,
-}: LogoProps) {
+export function Logo({ folded, size, className, onyxBranded }: LogoProps) {
   const resolvedSize = size ?? DEFAULT_LOGO_SIZE_PX;
   const { enterprise, logoUrl } = useSettings();
   const logoDisplayStyle = enterprise?.logo_display_style;
